Handle weather changes in generic form component
