+++ packages/website/src/routes/Surveys/New/Form.tsx
@@ -1,10 +1,9 @@
-import React, { useState, useCallback, ChangeEvent } from "react";
+import React, { useCallback } from "react";
 import { Grid, Collapse, IconButton } from "@material-ui/core";
 import Alert from "@material-ui/lab/Alert";
 import { useSelector, useDispatch } from "react-redux";
-import { SurveyData } from "../../../store/Survey/types";
+import { SurveyData, SurveyState } from "../../../store/Survey/types";
 import {
-  diveLocationSelector,
   surveyErrorSelector,
   surveyAddRequest,
 } from "../../../store/Survey/surveySlice";
@@ -12,26 +11,18 @@ import { userInfoSelector } from "../../../store/User/userSlice";
 import Form from "../../../common/SurveyForm";
 
 const SurveyForm = ({ reefId, changeTab }: SurveyFormProps) => {
-  const diveLocation = useSelector(diveLocationSelector);
   const user = useSelector(userInfoSelector);
-  const [weather, setWeather] = useState<SurveyData["weatherConditions"]>(
-    "calm"
-  );
   const surveyError = useSelector(surveyErrorSelector);
 
   const dispatch = useDispatch();
 
-  const handleWeatherChange = (event: ChangeEvent<{ value: unknown }>) => {
-    setWeather(event.target.value as SurveyData["weatherConditions"]);
-  };
-
   const onSubmit = useCallback(
-    (data: any) => {
-      const diveDateTime = new Date(
-        `${data.diveDate}, ${data.diveTime}`
-      ).toISOString();
-      const weatherConditions = weather;
-      const { comments } = data;
+    (
+      diveDateTime: string,
+      diveLocation: SurveyState["diveLocation"],
+      weatherConditions: SurveyData["weatherConditions"],
+      comments: string
+    ) => {
       const surveyData: SurveyData = {
         reef: reefId,
         diveDate: diveDateTime,
@@ -48,7 +39,7 @@ const SurveyForm = ({ reefId, changeTab }: SurveyFormProps) => {
         })
       );
     },
-    [dispatch, weather, diveLocation, changeTab, user, reefId]
+    [dispatch, changeTab, reefId, user]
   );
 
   return (
@@ -65,12 +56,7 @@ const SurveyForm = ({ reefId, changeTab }: SurveyFormProps) => {
           </Alert>
         </Collapse>
       </Grid>
-      <Form
-        reefId={reefId}
-        onSubmit={onSubmit}
-        handleWeatherChange={handleWeatherChange}
-        weather={weather}
-      />
+      <Form reefId={reefId} onSubmit={onSubmit} />
     </>
   );
 };
